pyproject.toml
[project]
name = "PA2025"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "accelerate==1.6.0",
    "beautifulsoup4==4.13.3",
    "black==25.1.0",
    "diffusers==0.33.1",
    "fastapi>=0.115.12",
    "fiftyone==1.5.2",
    "flask==3.1.0",
    "matplotlib>=3.10.3",
    "maturin>=1.8.7",
    "numpy==2.3.0",
    "pillow==11.2.1",
    "playwright==1.51.0",
    "pydantic>=2.11.5",
    "pymongo==4.9.2",
    "python-dotenv==1.1.0",
    "requests==2.32.3",
    "streamlit==1.45.0",
    "torch==2.7.0",
    "tqdm==4.67.1",
    "transformers==4.51.3",
]
